refactor drum switches to be reusable and update sheet music search route to ILIKE
