+++ server/routes/sheetMusic.router.js
@@ -24,10 +24,10 @@ router.get('/search/:instrument/:difficulty/:name', rejectUnauthenticated, (req,
     if (difficulty === '*') { difficulty = null }
     if (instrument === '*') { instrument = null }
     const queryString = `SELECT * from sheet_music where
-                        ($1::text is NULL or "name" = $1) and
+                        ($1::text is NULL or "name" ILIKE $1) and
                         ($2::difficulty is NULL or "difficulty" = $2) and
                         ($3::instrument is NULL or "instrument" = $3);`;
-    const queryValues = [ name, difficulty, instrument]
+    const queryValues = [ `%${name}%`, difficulty, instrument]
     pool.query(queryString, queryValues)
         .then((result) => {
             console.log(result.rows);
