+++ src/components/PanSecond/PanSecond.jsx
@@ -1,9 +1,6 @@
 import React, { Component } from 'react';
-// nav bar import
-import Switches from './Switches';
-// css import
+import Switches from '../PanSwitches/PanSwitches'
 import './PanSecond.css';
-// svg imports
 import ANotes from './ANotes';
 import BNotes from './BNotes';
 import BbNotes from './BbNotes';
